Added actors
